+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### 1.0.1
 - `MarkerTheme.corners()` - piece marker that draws triangles in the corners of the square.
+- Fixed piece selectors not closing on premoves.
+- `BoardController.promotionBehaviour` - this allows different behaviours to be specified: always show the selector (default), always auto select the best piece, or auto select only for premoves.
+- `BoardController.pieceHierarchy` - allows the relative hierarchy of pieces to be specified, for use in auto promotion and ordering the pieces in the selector.
 
 ### 1.0.0
 - Major refactor with improved stability, better performance and better organised code.